refactor(tui): キーバインド定義を keys/ に切り出す
app.py の約半分をキーバインド定義が占めていたため、通常モード (normal.py)、
modal 表示中と各モード (modal.py)、モード判定と共有ヘルパ (shared.py) に分ける。

キーバインドが共有していた 3 つのクロージャは state を引数に取る純関数にした。
これに伴い、キーバインドから呼ぶ選択肢構築とプロジェクト切替を app.py に
残すと相互 import になるため、choices.py と project.py へ移す。

キー定義 94 件とハンドラ本体は移動のみで変更していない。
